Split file I/O routines into fileio.c
