+++ ee.c
@@ -112,6 +112,7 @@ nl_catd catalog;
 #include "text.h"
 
 #include "buffer.h"
+#include "fileio.h"
 struct text *first_line;
 struct text *dlt_line;
 struct text *curr_line;
@@ -253,14 +254,11 @@ void goto_line(char *cmd_str);
 void midscreen(int line, ee_char *pnt);
 void get_options(int numargs, char *arguments[]);
 void check_fp(void);
-void get_file(char *file_name);
-void get_line(int length, ee_char *in_string, int *append);
 void draw_screen(void);
 void finish(void);
 int quit(int noverify);
 void edit_abort(int arg);
 void delete_text(void);
-int write_file(char *file_name, int warn_if_exists);
 int search(int display_message);
 void search_prompt(void);
 void del_char(void);
@@ -2236,146 +2234,7 @@ check_fp(void)
 	wrefresh(text_win);
 }
 
-/* read specified file into current buffer	*/
-void 
-get_file(char *file_name)
-{
-	int can_read;		/* file has at least one character	*/
-	int length;		/* length of line read by read		*/
-	int append;		/* should text be appended to current line */
-	struct text *temp_line;
-	char ro_flag = FALSE;
 
-	if (recv_file)		/* if reading a file			*/
-	{
-		wmove(com_win, 0, 0);
-		wclrtoeol(com_win);
-		wprintw(com_win, reading_file_msg, file_name);
-		if (access(file_name, 2))	/* check permission to write */
-		{
-			if ((errno == ENOTDIR) || (errno == EACCES) || (errno == EROFS) || (errno == ETXTBSY) || (errno == EFAULT))
-			{
-				wprintw(com_win, "%s", read_only_msg);
-				ro_flag = TRUE;
-			}
-		}
-		wrefresh(com_win);
-	}
-	if (curr_line->line_length > 1)	/* if current line is not blank	*/
-	{
-		insert_line(FALSE);
-		left(FALSE);
-		append = FALSE;
-	}
-	else
-		append = TRUE;
-	can_read = FALSE;		/* test if file has any characters  */
-        ee_char wbuf[513];
-        while (((length = read(get_fd, in_string, 512)) != 0) && (length != -1))
-        {
-                can_read = TRUE;  /* if set file has at least 1 character   */
-                in_string[length] = '\0';
-                int wlen = mbstowcs(wbuf, in_string, 513);
-                if (wlen < 0)
-                        wlen = 0;
-                get_line(wlen, wbuf, &append);
-        }
-	if ((can_read) && (curr_line->line_length == 1))
-	{
-		temp_line = curr_line->prev_line;
-		temp_line->next_line = curr_line->next_line;
-		if (temp_line->next_line != NULL)
-			temp_line->next_line->prev_line = temp_line;
-		if (curr_line->line != NULL)
-			free(curr_line->line);
-		free(curr_line);
-		curr_line = temp_line;
-	}
-	if (input_file)	/* if this is the file to be edited display number of lines	*/
-	{
-		wmove(com_win, 0, 0);
-		wclrtoeol(com_win);
-		wprintw(com_win, file_read_lines_msg, in_file_name, curr_line->line_number);
-		if (ro_flag)
-			wprintw(com_win, "%s", read_only_msg);
-		wrefresh(com_win);
-	}
-	else if (can_read)	/* not input_file and file is non-zero size */
-		text_changes = TRUE;
-
-	if (recv_file)		/* if reading a file			*/
-	{
-		in = EOF;
-	}
-}
-
-/* read string and split into lines */
-void 
-get_line(int length, ee_char *in_string, int *append)
-{
-	ee_char *str1;
-	ee_char *str2;
-	int num;		/* offset from start of string		*/
-	int char_count;		/* length of new line (or added portion	*/
-	int temp_counter;	/* temporary counter value		*/
-	struct text *tline;	/* temporary pointer to new line	*/
-	int first_time;		/* if TRUE, the first time through the loop */
-
-	str2 = in_string;
-	num = 0;
-	first_time = TRUE;
-	while (num < length)
-	{
-		if (!first_time)
-		{
-			if (num < length)
-			{
-				str2++;
-				num++;
-			}
-		}
-		else
-			first_time = FALSE;
-		str1 = str2;
-		char_count = 1;
-		/* find end of line	*/
-		while ((*str2 != '\n') && (num < length))
-		{
-			str2++;
-			num++;
-			char_count++;
-		}
-		if (!(*append))	/* if not append to current line, insert new one */
-		{
-			tline = txtalloc();	/* allocate data structure for next line */
-			tline->line_number = curr_line->line_number + 1;
-			tline->next_line = curr_line->next_line;
-			tline->prev_line = curr_line;
-			curr_line->next_line = tline;
-			if (tline->next_line != NULL)
-				tline->next_line->prev_line = tline;
-			curr_line = tline;
-                        curr_line->line = point = malloc(char_count * sizeof(ee_char));
-			curr_line->line_length = char_count;
-			curr_line->max_length = char_count;
-		}
-		else
-		{
-			point = resiz_line(char_count, curr_line, curr_line->line_length); 
-			curr_line->line_length += (char_count - 1);
-		}
-		for (temp_counter = 1; temp_counter < char_count; temp_counter++)
-		{
-			*point = *str1;
-			point++;
-			str1++;
-		}
-		*point = '\0';
-		*append = FALSE;
-		if ((num == length) && (*str2 != '\n'))
-			*append = TRUE;
-	}
-}
 
 void 
 draw_screen()		/* redraw the screen from current postion	*/
@@ -2505,76 +2364,6 @@ delete_text(void)
 	position = 1;
 }
 
-int 
-write_file(char *file_name, int warn_if_exists)
-{
-	char cr;
-	char *tmp_point;
-	struct text *out_line;
-        int lines, charac;
-        int write_flag = TRUE;
-
-	charac = lines = 0;
-	if (warn_if_exists &&
-	    ((in_file_name == NULL) || strcmp(in_file_name, file_name)))
-	{
-		if ((temp_fp = fopen(file_name, "r")))
-		{
-			tmp_point = get_string(file_exists_prompt, TRUE);
-                        if (towupper(*tmp_point) == towupper(*yes_char))
-				write_flag = TRUE;
-			else 
-				write_flag = FALSE;
-			fclose(temp_fp);
-			free(tmp_point);
-		}
-	}
-
-	clear_com_win = TRUE;
-
-	if (write_flag)
-	{
-		if ((temp_fp = fopen(file_name, "w")) == NULL)
-		{
-			clear_com_win = TRUE;
-			wmove(com_win,0,0);
-			wclrtoeol(com_win);
-			wprintw(com_win, create_file_fail_msg, file_name);
-			wrefresh(com_win);
-			return(FALSE);
-		}
-		else
-		{
-			wmove(com_win,0,0);
-			wclrtoeol(com_win);
-			wprintw(com_win, writing_file_msg, file_name);
-			wrefresh(com_win);
-			cr = '\n';
-                        out_line = first_line;
-                        while (out_line != NULL)
-                        {
-                                size_t bytes;
-                                char mbbuf[4096];
-                                bytes = wcstombs(mbbuf, out_line->line, sizeof(mbbuf));
-                                if (bytes == (size_t)-1)
-                                        bytes = 0;
-                                fwrite(mbbuf, 1, bytes, temp_fp);
-                                charac += bytes;
-                                out_line = out_line->next_line;
-                                putc(cr, temp_fp);
-                                lines++;
-                        }
-			fclose(temp_fp);
-			wmove(com_win,0,0);
-			wclrtoeol(com_win);
-			wprintw(com_win, file_written_msg, file_name, lines, charac);
-			wrefresh(com_win);
-			return(TRUE);
-		}
-	}
-	else
-		return(FALSE);
-}
 
 /* search for string in srch_str	*/
 int
